pyproject.toml
[build-system]
requires = [

  "hatchling>=1.6",
  "mypy",
  "mypy_extensions",
  "types-setuptools",
  "types-tabulate",
  "types-tqdm",
  ## mypy requires type stubs in build environment
  "typing_extensions",
  ]
  
build-backend = "hatchling.build"


[tool.hatch.build.targets.sdist.hooks.mypyc]
dependencies = ["hatch-mypyc"]

[build.targets.sdist.hooks.mypyc]
include = ["src"]
mypy-args = [
  "--verbose",
  "--install-types",
  "--non-interactive",
  "--config-file=mypy.ini",
]

[tool.hatch.build.wheel.sdist.hooks.mypyc]
dependencies = ["hatch-mypyc"]

[build.targets.wheel.hooks.mypyc]
mypy-args = [
  "--config-file=mypy.ini",
]


[project]
name = "torchsift"
dynamic = ["version"]
description = "TF-VPR"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE.txt"}
keywords = ["SIFT", "CV", "torch"]
authors = [
  {email = "zc2309@nyu.edu"},
  {name = "Zegang Cheng"}
]
maintainers = [
  {name = "Zegang Cheng", email = "zc2309@nyu.edu"}
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "License :: OSI Approved :: MIT License",
]

dependencies = [
  "pandas",
  "opencv_python>=4.7.0",
  "torch==1.13.1",
  "torchvision==0.14.1",
  "Pillow",
  "types-Pillow",
]


[project.urls]
homepage = "https://github.com/chengzegang/TorchSIFT"
documentation = "https://github.com/chengzegang/TorchSIFT"
repository = "https://github.com/chengzegang/TorchSIFT"


[project.optional-dependencies]
test = [
  "coverage[toml]",
  'pytest',
  "pytest-cov",
  "pytest-sugar"
]

dev = [
  'black',
  'flake8',
  'isort',
]

[tool.hatch.build.target.sdist]

exclude = ['tests']

[tool.hatch.build.targets.wheel]
packages = ["src"]


[tool.hatch.version]
path = "src/torchsift/__about__.py"


[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
allow_redefinition = true
[[tool.mypy.overrides]]
module = [
    "torch.*",
    "torchvision.*",
    "pykeops.*",
    "scipy.*",
    "matplotlib.*",
    "ot.*",
    "kornia.*",
]
ignore_missing_imports = true